requirements.txt
asgiref==3.5.1
Django==3.2.6
pytz==2022.1
sqlparse==0.4.2
psycopg2-binary==2.9.3
djangorestframework==3.12.4